requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pdfminer.six>=20221105
python-docx>=1.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pydantic>=2.7.1
langchain>=0.2.0
openai>=1.30.1
chroma-embedding-functions>=0.3.2
